# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Control API
  version: 1.0.0
  description: 单 Deployment Control 首版合同。/auth/v1 供 Nexus Web 同源访问；internal 路由只供受信下游服务调用。
servers:
  - url: /
paths:
...
  /auth/v1/login:
    post:
      operationId: webLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebLoginInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
  /auth/v1/logout:
    post:
      operationId: webLogout
      security:
        - CookieSession: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Error"
  /auth/v1/setup:
    post:
      operationId: webSetup
      security:
        - SetupToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupOwnerInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/v1/profile:
    patch:
      operationId: webUpdateProfile
      security:
        - CookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /auth/v1/profile/password:
    post:
      operationId: webChangePassword
      security:
        - CookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/v1/profile/password/receipt:
    get:
      operationId: webPasswordReceipt
      security:
        - CookieSession: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/control/v1/setup/owner:
    post:
      operationId: setupOwner
      security:
        - SetupToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupOwnerInput"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/control/v1/internal/state:
    get:
      operationId: internalState
      x-internal: true
      security:
        - ServiceToken: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/control/v1/internal/principals/exchange:
    post:
      operationId: exchangePrincipal
      x-internal: true
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [session_token, audience]
              properties:
                session_token:
                  type: string
                audience:
                  type: string
                  description: 配置的 Runtime audience，或固定的 nexus-relay-user。
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /api/control/v1/internal/identity-invalidations:
    get:
      operationId: listIdentityInvalidations
      x-internal: true
      security:
        - ServiceToken: []
      parameters:
        - in: query
          name: after
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 256
            default: 256
      responses:
        "200":
          description: 有序身份失效事件；next_cursor 等于最后一个 event_id。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [events, next_cursor]
                        properties:
                          events:
                            type: array
                            items:
                              $ref: "#/components/schemas/IdentityInvalidation"
                          next_cursor:
                            type: integer
                            format: int64
        "400":
          $ref: "#/components/responses/Error"
...
  /api/control/v1/internal/organizations/members/verify:
    post:
      operationId: verifyOrganizationMembers
      description: 供受信 Nexus Gateway 在建群前确认所有真人目标都属于指定 Deployment 下的同一 active Organization。
      x-internal: true
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deployment_id, organization_id, user_ids]
              properties:
                deployment_id: {type: string}
                organization_id: {type: string}
                user_ids:
                  type: array
                  maxItems: 100
                  items: {type: string}
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Error"
  /api/control/v1/internal/agents/verify:
    post:
      operationId: verifyOwnedAgents
      description: 供受信 Nexus Gateway 验证一组 active Agent 均属于指定组织和真人，并返回可投影到 Relay 的公开身份。
      x-internal: true
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deployment_id, organization_id, owner_user_id, agent_ids]
              properties:
                deployment_id: {type: string}
                organization_id: {type: string}
                owner_user_id: {type: string}
                agent_ids:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: {type: string}
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    NodeCredential:
      type: http
      scheme: bearer
      description: 注册时由设备安全保存的独立随机凭据，不是浏览器 Session 或服务 token。
    ServiceToken:
      type: http
      scheme: bearer
    SetupToken:
      type: http
      scheme: bearer
    CookieSession:
      type: apiKey
      in: cookie
      name: nexus_session
  parameters:
...
    RequestID:
      name: request_id
      in: query
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    SetupOwnerInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
        display_name:
          type: string
        password:
          type: string
          minLength: 8
        deployment_name:
          type: string
          maxLength: 128
        organization_name:
          type: string
          maxLength: 128
    WebLoginInput:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 1
          maxLength: 1024
    UpdateProfileInput:
      type: object
      additionalProperties: false
      required: [avatar]
      properties:
        avatar:
          type: string
          maxLength: 255
...
    ChangePasswordInput:
      type: object
      additionalProperties: false
      required: [request_id, current_password, new_password]
      properties:
        request_id:
          type: string
          minLength: 8
          maxLength: 128
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8
    IdentityInvalidation:
      type: object
      required: [event_id, deployment_id, user_id, reason, created_at]
      properties:
        organization_id:
          type: string
          description: 组织成员撤权的精确作用域。
        membership_revoked:
          type: boolean
          description: 已提交的组织撤权事实；Relay 必须与消费游标同事务处理，不得仅看当前成员状态推断历史。
        event_id:
          type: integer
          format: int64
        deployment_id:
          type: string
        user_id:
          type: string
        session_id:
          type: string
          description: 仅 session_revoked 事件携带；普通值撤销浏览器 Session 及其派生 Node，node:<node_id> 只撤销精确 Node。
        reason:
          type: string
          enum: [principal_changed, profile_changed, session_revoked, entitlement_changed, organization_changed]
        created_at:
          type: string
          format: date-time
    Envelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        data: {}
        details:
          type: object
          additionalProperties: true
  responses:
    Success:
      description: 请求成功，业务载荷位于 data。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 请求失败。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"